Fix subcommand append when module is a callable function
Directly passing module as a click callback yields a weird behavior
when the module is a function which accepts yper arguments and options
In this case, their default values are neither checked nor assigned.
To properly handle, construct typer app and append as a command.
Actually this is identical to the original implementation
except registering origins!

Additionally, we have to reassign the help message of the original
command. I am not sure if direct manipulation of attribute of click is a
good idea...